HistScanner added some CFD timing.
Added branches for baseline, timestamp (timeStampNs), cfdBin, and event
time from the timestamp and cfd (timeCfdNs).

Updated HistScanner and RootScanner to handle double ranges.

Currently, the HistScanner does not properly plot timeStampNs and
timeCfdNs as they are doubles.


Former-commit-id: 6978389473dd3d629206904fe384f89468536525

+++ Scan/util/source/RootScanner.cpp
@@ -35,10 +35,10 @@ void RootScanner::IdleTask() {
 void RootScanner::ResetZoom(TVirtualPad *pad /*= gPad*/) {
 	AxisInfo* padZoomInfo = &zoomInfo_[pad];
 	for (int i=0;i<numAxes_;i++) {
-		padZoomInfo->limitMin[i] = std::numeric_limits<float>::max();
-		padZoomInfo->limitMax[i] = std::numeric_limits<float>::min();
-		padZoomInfo->rangeUserMin[i] = std::numeric_limits<float>::max();
-		padZoomInfo->rangeUserMax[i] = std::numeric_limits<float>::min();
+		padZoomInfo->limitMin[i] = std::numeric_limits<double>::max();
+		padZoomInfo->limitMax[i] = std::numeric_limits<double>::min();
+		padZoomInfo->rangeUserMin[i] = std::numeric_limits<double>::max();
+		padZoomInfo->rangeUserMax[i] = std::numeric_limits<double>::min();
 	}
 	padZoomInfo->reset = true;
 }
